Fix -Wsign-compare warnings across the codebase
Fix all 15 signed/unsigned comparison warnings triggered by
-Wsign-compare:

  pahole.c: use size_t for ARRAY_SIZE() loop iterators in
  init_btf_features(), find_btf_feature(),
  show_supported_btf_features() and btf_features__enable_default(),
  moving declarations into the for statements.

  btf_encoder.c: use size_t for ARRAY_SIZE() loops in
  should_skip_decl() and btf_encoder__should_skip_kfunc(), use
  size_t for the offset in is_sym_kfunc_set(), and unsigned int
  for the loop counter matching ranges_cnt.

  btf_loader.c: cast bitfield arithmetic to size_t in ternary
  where the other branch is pos->byte_size (size_t).

  dwarf_loader.c: cast int8_t bitfield_offset to uint32_t before
  comparing against uint32_t bit_size — by this point the value
  has already been corrected to non-negative.

  dwarves.c: use uint16_t for byte_hole_size to match the int
  hole field it compares against, and add an id >= 0 guard before
  the ARRAY_SIZE() comparison in lang__int2str().

  dwarves_fprintf.c: use uint64_t for loop variable matching
  attributes->cnt type.

  dwarves_reorganize.c: use int for remainder/inc (bounded by
  addr_size), and cast byte_size comparison to size_t where the
  hole value is known non-negative.

Fixes: 7bc9b9975545ab53 ("pahole: Add --btf_features support")
Link: https://lore.kernel.org/dwarves/20260729190733.72876-1-acme@kernel.org/

+++ btf_loader.c
@@ -744,7 +744,7 @@ static int class__fixup_btf_bitfields(const struct conf_load *conf, struct tag *
 		 */
 		smallest_offset = pos->byte_offset;
 		smallest_offset += pos->bitfield_size ?
-			(pos->bitfield_offset + pos->bitfield_size + 7) / 8 :
+			(size_t)(pos->bitfield_offset + pos->bitfield_size + 7) / 8 :
 			pos->byte_size;
 	}
